Adding Path Verification to Watch Task + Updated CSS

+++ docs/watch.rb
@@ -12,42 +12,46 @@ path_from_dist = data_hash['path_from_dist'];
 path_from_css = path_from_dist + 'css'
 path_to_dist_css = './assets/dist/css'
 
-puts "Watching -- " + path_from_css
+if path_from_dist.eql? "<path to wwwroot dist>"
+  raise ArgumentError, 'You need to configure your path_from_dist variable, point this to the location of your Wildernesslabs wwwroot dist directory and try running this task again.'
+else
+  puts "Watching -- " + path_from_css
 
-Filewatcher.new([path_from_css]).watch do |filename, event|
+  Filewatcher.new([path_from_css]).watch do |filename, event|
 
-  # specify files which should not be copied
-  puts "Copying files from WildernessLabs CSS to Documentation site --------------- "
+    # specify files which should not be copied
+    puts "Copying files from WildernessLabs CSS to Documentation site --------------- "
 
-  from_dir = path_from_css
-  to_dir = path_to_dist_css
+    from_dir = path_from_css
+    to_dir = path_to_dist_css
 
-  contains = Dir.new(from_dir).entries
+    contains = Dir.new(from_dir).entries
 
-  def copy_with_path(src, dst)
-    FileUtils.mkdir_p(File.dirname(dst))
-    FileUtils.cp(src, dst)
+    def copy_with_path(src, dst)
+      FileUtils.mkdir_p(File.dirname(dst))
+      FileUtils.cp(src, dst)
+    end
+
+    puts "Emptying directory target --------------- "
+    FileUtils.rm_rf(to_dir)
+
+    puts "Copying new files --------------- "
+    FileUtils.cp_r(from_dir, to_dir);
+    puts "Copying CSS Complete --------------- "
+
+    puts "Updating file paths in CSS --------------- "
+    file_names = [path_to_dist_css+'/main.css']
+
+    file_names.each do |file_name|
+      text = File.read(file_name)
+      new_contents = text.gsub("../../assets/", "/assets/dist/assets/")
+
+      # To write changes to the file, use:
+      File.open(file_name, "w") {|file| file.puts new_contents }
+    end
+
+    puts "Path Updating Complete --------------- "
+    puts "Watching " + path_from_css
+    puts "-"
   end
-
-  puts "Emptying directory target --------------- "
-  FileUtils.rm_rf(to_dir)
-
-  puts "Copying new files --------------- "
-  FileUtils.cp_r(from_dir, to_dir);
-  puts "Copying CSS Complete --------------- "
-
-  puts "Updating file paths in CSS --------------- "
-  file_names = [path_to_dist_css+'/main.css']
-
-  file_names.each do |file_name|
-    text = File.read(file_name)
-    new_contents = text.gsub("../../assets/", "/assets/dist/assets/")
-
-    # To write changes to the file, use:
-    File.open(file_name, "w") {|file| file.puts new_contents }
-  end
-
-  puts "Path Updating Complete --------------- "
-  puts "Watching " + path_from_css
-  puts "-"
 end
